Considering AST interpreter done
There are some minor differences when it comes to error messages and
some of the additional feature (such as class methods) make the jlox
suite fail.

It's surprisingly much slower than jlox.. I'm not sure whether I
should invest more time into optimizing the tree walking interpreter
or just move on to the bytecode VM as that's the approach I'd chose
anyways when wanting a fast scripting language...

Let's not dwell too much on the AST interpreter for now and move on.
I can always return to this one.